+++ sdk/nodejs/types/output.ts
@@ -2658,6 +2658,95 @@ export namespace azure {
          */
         healthCheckType?: string;
     }
+
+    export interface OceanVirtualNodeGroupAutoscale {
+        autoscaleHeadroom?: outputs.azure.OceanVirtualNodeGroupAutoscaleAutoscaleHeadroom;
+    }
+
+    export interface OceanVirtualNodeGroupAutoscaleAutoscaleHeadroom {
+        /**
+         * Configure the number of CPUs to allocate for the headroom. CPUs are denoted in millicores, where 1000 millicores = 1 vCPU.
+         */
+        cpuPerUnit?: number;
+        /**
+         * How many GPU cores should be allocated for headroom unit.
+         */
+        gpuPerUnit?: number;
+        /**
+         * Configure the amount of memory (MiB) to allocate the headroom.
+         */
+        memoryPerUnit?: number;
+        /**
+         * The number of headroom units to maintain, where each unit has the defined CPU, memory and GPU.
+         */
+        numOfUnits: number;
+    }
+
+    export interface OceanVirtualNodeGroupLabel {
+        /**
+         * Tag Key for Vms in the cluster.
+         */
+        key: string;
+        /**
+         * Tag Value for VMs in the cluster.
+         */
+        value?: string;
+    }
+
+    export interface OceanVirtualNodeGroupLaunchSpecification {
+        /**
+         * Specify OS disk specification other than default.
+         */
+        osDisk?: outputs.azure.OceanVirtualNodeGroupLaunchSpecificationOsDisk;
+        /**
+         * Additional key-value pairs to be used to tag the VMs in the virtual node group.
+         */
+        tags?: outputs.azure.OceanVirtualNodeGroupLaunchSpecificationTag[];
+    }
+
+    export interface OceanVirtualNodeGroupLaunchSpecificationOsDisk {
+        /**
+         * The size of the OS disk in GB, Required if dataDisks is specified.
+         */
+        sizeGb: number;
+        /**
+         * The type of the OS disk. Valid values: `"Standard_LRS"`, `"Premium_LRS"`, `"StandardSSD_LRS"`.
+         */
+        type?: string;
+    }
+
+    export interface OceanVirtualNodeGroupLaunchSpecificationTag {
+        /**
+         * Tag Key for Vms in the cluster.
+         */
+        key?: string;
+        /**
+         * Tag Value for VMs in the cluster.
+         */
+        value?: string;
+    }
+
+    export interface OceanVirtualNodeGroupResourceLimit {
+        /**
+         * Option to set a maximum number of instances per virtual node group. If set, value must be greater than or equal to 0.
+         */
+        maxInstanceCount?: number;
+    }
+
+    export interface OceanVirtualNodeGroupTaint {
+        /**
+         * The effect of the taint. Valid values: `"NoSchedule"`, `"PreferNoSchedule"`, `"NoExecute"`, `"PreferNoExecute"`.
+         */
+        effect: string;
+        /**
+         * Tag Key for Vms in the cluster.
+         */
+        key: string;
+        /**
+         * Tag Value for VMs in the cluster.
+         */
+        value: string;
+    }
 }
 
 export namespace ecs {