+++ sdk/python/pulumi_spotinst/azure/outputs.py
@@ -29,6 +29,14 @@ __all__ = [
     'ElastigroupStrategy',
     'ElastigroupUpdatePolicy',
     'ElastigroupUpdatePolicyRollConfig',
+    'OceanVirtualNodeGroupAutoscale',
+    'OceanVirtualNodeGroupAutoscaleAutoscaleHeadroom',
+    'OceanVirtualNodeGroupLabel',
+    'OceanVirtualNodeGroupLaunchSpecification',
+    'OceanVirtualNodeGroupLaunchSpecificationOsDisk',
+    'OceanVirtualNodeGroupLaunchSpecificationTag',
+    'OceanVirtualNodeGroupResourceLimit',
+    'OceanVirtualNodeGroupTaint',
 ]
 
 @pulumi.output_type
@@ -1200,6 +1208,278 @@ class ElastigroupUpdatePolicyRollConfig(dict):
         Sets the health check type to use. Valid values: `"INSTANCE_STATE"`, `"NONE"`.
         """
         return pulumi.get(self, "health_check_type")
+
+    def _translate_property(self, prop):
+        return _tables.CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop
+
+
+@pulumi.output_type
+class OceanVirtualNodeGroupAutoscale(dict):
+    def __init__(__self__, *,
+                 autoscale_headroom: Optional['outputs.OceanVirtualNodeGroupAutoscaleAutoscaleHeadroom'] = None):
+        if autoscale_headroom is not None:
+            pulumi.set(__self__, "autoscale_headroom", autoscale_headroom)
+
+    @property
+    @pulumi.getter(name="autoscaleHeadroom")
+    def autoscale_headroom(self) -> Optional['outputs.OceanVirtualNodeGroupAutoscaleAutoscaleHeadroom']:
+        return pulumi.get(self, "autoscale_headroom")
+
+    def _translate_property(self, prop):
+        return _tables.CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop
+
+
+@pulumi.output_type
+class OceanVirtualNodeGroupAutoscaleAutoscaleHeadroom(dict):
+    def __init__(__self__, *,
+                 num_of_units: int,
+                 cpu_per_unit: Optional[int] = None,
+                 gpu_per_unit: Optional[int] = None,
+                 memory_per_unit: Optional[int] = None):
+        """
+        :param int num_of_units: The number of headroom units to maintain, where each unit has the defined CPU, memory and GPU.
+        :param int cpu_per_unit: Configure the number of CPUs to allocate for the headroom. CPUs are denoted in millicores, where 1000 millicores = 1 vCPU.
+        :param int gpu_per_unit: How many GPU cores should be allocated for headroom unit.
+        :param int memory_per_unit: Configure the amount of memory (MiB) to allocate the headroom.
+        """
+        pulumi.set(__self__, "num_of_units", num_of_units)
+        if cpu_per_unit is not None:
+            pulumi.set(__self__, "cpu_per_unit", cpu_per_unit)
+        if gpu_per_unit is not None:
+            pulumi.set(__self__, "gpu_per_unit", gpu_per_unit)
+        if memory_per_unit is not None:
+            pulumi.set(__self__, "memory_per_unit", memory_per_unit)
+
+    @property
+    @pulumi.getter(name="numOfUnits")
+    def num_of_units(self) -> int:
+        """
+        The number of headroom units to maintain, where each unit has the defined CPU, memory and GPU.
+        """
+        return pulumi.get(self, "num_of_units")
+
+    @property
+    @pulumi.getter(name="cpuPerUnit")
+    def cpu_per_unit(self) -> Optional[int]:
+        """
+        Configure the number of CPUs to allocate for the headroom. CPUs are denoted in millicores, where 1000 millicores = 1 vCPU.
+        """
+        return pulumi.get(self, "cpu_per_unit")
+
+    @property
+    @pulumi.getter(name="gpuPerUnit")
+    def gpu_per_unit(self) -> Optional[int]:
+        """
+        How many GPU cores should be allocated for headroom unit.
+        """
+        return pulumi.get(self, "gpu_per_unit")
+
+    @property
+    @pulumi.getter(name="memoryPerUnit")
+    def memory_per_unit(self) -> Optional[int]:
+        """
+        Configure the amount of memory (MiB) to allocate the headroom.
+        """
+        return pulumi.get(self, "memory_per_unit")
+
+    def _translate_property(self, prop):
+        return _tables.CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop
+
+
+@pulumi.output_type
+class OceanVirtualNodeGroupLabel(dict):
+    def __init__(__self__, *,
+                 key: str,
+                 value: Optional[str] = None):
+        """
+        :param str key: Tag Key for Vms in the cluster.
+        :param str value: Tag Value for VMs in the cluster.
+        """
+        pulumi.set(__self__, "key", key)
+        if value is not None:
+            pulumi.set(__self__, "value", value)
+
+    @property
+    @pulumi.getter
+    def key(self) -> str:
+        """
+        Tag Key for Vms in the cluster.
+        """
+        return pulumi.get(self, "key")
+
+    @property
+    @pulumi.getter
+    def value(self) -> Optional[str]:
+        """
+        Tag Value for VMs in the cluster.
+        """
+        return pulumi.get(self, "value")
+
+    def _translate_property(self, prop):
+        return _tables.CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop
+
+
+@pulumi.output_type
+class OceanVirtualNodeGroupLaunchSpecification(dict):
+    def __init__(__self__, *,
+                 os_disk: Optional['outputs.OceanVirtualNodeGroupLaunchSpecificationOsDisk'] = None,
+                 tags: Optional[Sequence['outputs.OceanVirtualNodeGroupLaunchSpecificationTag']] = None):
+        """
+        :param 'OceanVirtualNodeGroupLaunchSpecificationOsDiskArgs' os_disk: Specify OS disk specification other than default.
+        :param Sequence['OceanVirtualNodeGroupLaunchSpecificationTagArgs'] tags: Additional key-value pairs to be used to tag the VMs in the virtual node group.
+        """
+        if os_disk is not None:
+            pulumi.set(__self__, "os_disk", os_disk)
+        if tags is not None:
+            pulumi.set(__self__, "tags", tags)
+
+    @property
+    @pulumi.getter(name="osDisk")
+    def os_disk(self) -> Optional['outputs.OceanVirtualNodeGroupLaunchSpecificationOsDisk']:
+        """
+        Specify OS disk specification other than default.
+        """
+        return pulumi.get(self, "os_disk")
+
+    @property
+    @pulumi.getter
+    def tags(self) -> Optional[Sequence['outputs.OceanVirtualNodeGroupLaunchSpecificationTag']]:
+        """
+        Additional key-value pairs to be used to tag the VMs in the virtual node group.
+        """
+        return pulumi.get(self, "tags")
+
+    def _translate_property(self, prop):
+        return _tables.CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop
+
+
+@pulumi.output_type
+class OceanVirtualNodeGroupLaunchSpecificationOsDisk(dict):
+    def __init__(__self__, *,
+                 size_gb: int,
+                 type: Optional[str] = None):
+        """
+        :param int size_gb: The size of the OS disk in GB, Required if dataDisks is specified.
+        :param str type: The type of the OS disk. Valid values: `"Standard_LRS"`, `"Premium_LRS"`, `"StandardSSD_LRS"`.
+        """
+        pulumi.set(__self__, "size_gb", size_gb)
+        if type is not None:
+            pulumi.set(__self__, "type", type)
+
+    @property
+    @pulumi.getter(name="sizeGb")
+    def size_gb(self) -> int:
+        """
+        The size of the OS disk in GB, Required if dataDisks is specified.
+        """
+        return pulumi.get(self, "size_gb")
+
+    @property
+    @pulumi.getter
+    def type(self) -> Optional[str]:
+        """
+        The type of the OS disk. Valid values: `"Standard_LRS"`, `"Premium_LRS"`, `"StandardSSD_LRS"`.
+        """
+        return pulumi.get(self, "type")
+
+    def _translate_property(self, prop):
+        return _tables.CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop
+
+
+@pulumi.output_type
+class OceanVirtualNodeGroupLaunchSpecificationTag(dict):
+    def __init__(__self__, *,
+                 key: Optional[str] = None,
+                 value: Optional[str] = None):
+        """
+        :param str key: Tag Key for Vms in the cluster.
+        :param str value: Tag Value for VMs in the cluster.
+        """
+        if key is not None:
+            pulumi.set(__self__, "key", key)
+        if value is not None:
+            pulumi.set(__self__, "value", value)
+
+    @property
+    @pulumi.getter
+    def key(self) -> Optional[str]:
+        """
+        Tag Key for Vms in the cluster.
+        """
+        return pulumi.get(self, "key")
+
+    @property
+    @pulumi.getter
+    def value(self) -> Optional[str]:
+        """
+        Tag Value for VMs in the cluster.
+        """
+        return pulumi.get(self, "value")
+
+    def _translate_property(self, prop):
+        return _tables.CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop
+
+
+@pulumi.output_type
+class OceanVirtualNodeGroupResourceLimit(dict):
+    def __init__(__self__, *,
+                 max_instance_count: Optional[int] = None):
+        """
+        :param int max_instance_count: Option to set a maximum number of instances per virtual node group. If set, value must be greater than or equal to 0.
+        """
+        if max_instance_count is not None:
+            pulumi.set(__self__, "max_instance_count", max_instance_count)
+
+    @property
+    @pulumi.getter(name="maxInstanceCount")
+    def max_instance_count(self) -> Optional[int]:
+        """
+        Option to set a maximum number of instances per virtual node group. If set, value must be greater than or equal to 0.
+        """
+        return pulumi.get(self, "max_instance_count")
+
+    def _translate_property(self, prop):
+        return _tables.CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop
+
+
+@pulumi.output_type
+class OceanVirtualNodeGroupTaint(dict):
+    def __init__(__self__, *,
+                 effect: str,
+                 key: str,
+                 value: str):
+        """
+        :param str effect: The effect of the taint. Valid values: `"NoSchedule"`, `"PreferNoSchedule"`, `"NoExecute"`, `"PreferNoExecute"`.
+        :param str key: Tag Key for Vms in the cluster.
+        :param str value: Tag Value for VMs in the cluster.
+        """
+        pulumi.set(__self__, "effect", effect)
+        pulumi.set(__self__, "key", key)
+        pulumi.set(__self__, "value", value)
+
+    @property
+    @pulumi.getter
+    def effect(self) -> str:
+        """
+        The effect of the taint. Valid values: `"NoSchedule"`, `"PreferNoSchedule"`, `"NoExecute"`, `"PreferNoExecute"`.
+        """
+        return pulumi.get(self, "effect")
+
+    @property
+    @pulumi.getter
+    def key(self) -> str:
+        """
+        Tag Key for Vms in the cluster.
+        """
+        return pulumi.get(self, "key")
+
+    @property
+    @pulumi.getter
+    def value(self) -> str:
+        """
+        Tag Value for VMs in the cluster.
+        """
+        return pulumi.get(self, "value")
 
     def _translate_property(self, prop):
         return _tables.CAMEL_TO_SNAKE_CASE_TABLE.get(prop) or prop